refactor: split auth state and pkce modules
